--- src/app/dashboard/import-export/export.service.spec.ts
@@ -1,15 +1,0 @@
-import { TestBed, inject } from '@angular/core/testing';
-
-import { ExportService } from './export.service';
-
-describe('ExportService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [ExportService]
-    });
-  });
-
-  it('should be created', inject([ExportService], (service: ExportService) => {
-    expect(service).toBeTruthy();
-  }));
-});
