--- src/app/dashboard/import-export/import-modal/import-modal.component.css
@@ -1,8 +1,0 @@
-.btn{
-    padding: 6px;
-}
-
-
-.modal-dialog{
-    top: .75rem;
-}
