--- src/app/dashboard/import-export/import-modal/import-modal.component.html
@@ -1,45 +1,0 @@
-<div bsModal #importModal="bs-modal" class="modal fade" tabindex="-1" role="dialog" aria-labelledby="importModalLabel"
-    aria-hidden="true" [config]="{backdrop: 'static'}">
-    <div class="modal-dialog modal-md">
-        <div class="modal-content">
-            <div class="modal-header" *ngIf="!importInProgress">
-                <h4 class="mb-0">Choose .json or .gz File to be Imported</h4>
-                <button class="close pull-right" aria-label="Close" (click)="hideImportModal()">
-                    <span aria-hidden="true">&times;</span>
-                </button>
-            </div>
-            <div class="modal-body">
-                <p id="inputModalTitleHelp" class="mb-0 text-muted text-help">
-                    Assessments will be imported into working directory
-                </p>
-                <div class="row" *ngIf="!importInProgress">
-                    <div class="col-12 form-group">
-                        <input type="file" class="form-control-file" (change)="setImportFile($event)" accept=".json, .gz">
-                        <span class="alert-danger small" *ngIf="treasureFile == true">Invalid File Destination - This 
-                            is a treasure hunt opportunites file and must be imported from within a treasure hunt.</span>
-                        <span class="alert-danger small" *ngIf="validFile == false && treasureFile == false">Invalid File Type</span>
-                    </div>
-                </div>
-
-                <div class="row justify-content-center" *ngIf="importInProgress">
-                    <div class="col-md-10">
-                        <div class="main-content text-center">
-                            <div class="spinner">
-                                <div class="bounce1"></div>
-                                <div class="bounce2"></div>
-                                <div class="bounce3"></div>
-                            </div>
-                            <h1>Importing Data...</h1>
-                        </div>
-                    </div>
-                </div>
-
-            </div>
-            <div class="modal-footer" *ngIf="!importInProgress">
-                <button [disabled]="validFile !== true" class="btn btn-primary pull-right"
-                    (click)="importFile()">Import</button>
-                <button class="btn btn-secondary pull-left" (click)="hideImportModal()">Cancel</button>
-            </div>
-        </div>
-    </div>
-</div>
